removed unnecessary casting

+++ source/scenes/BossSelectScene.cpp
@@ -267,7 +267,7 @@ void BossSelectScene::configureLockButton() {
  */
 void BossSelectScene::slideTo(int newIndex) {
     if (_isAnimating) return;
-    if (newIndex < 0 || newIndex >= (int)_bossCards.size()) return;
+    if (newIndex < 0 || newIndex >= _bossCards.size()) return;
 
     _isAnimating = true;
